TestHarness - Layout of visual control in display-host.

+++ Source/Open.Core/JavaScript/Open.Core/Models/Size.cs
@@ -4,13 +4,13 @@
     public class Size
     {
         #region Head
-        private readonly double width;
-        private readonly double height;
+        private readonly int width;
+        private readonly int height;
 
         /// <summary>Constructor.</summary>
         /// <param name="width">The pixel width of the element.</param>
         /// <param name="height">The pixel height of the element.</param>
-        public Size(double width, double height)
+        public Size(int width, int height)
         {
             this.width = width;
             this.height = height;
@@ -19,10 +19,10 @@
 
         #region Properties
         /// <summary>Gets or sets the pixel width of the element.</summary>
-        public double Width { get { return width; } }
+        public int Width { get { return width; } }
 
         /// <summary>Gets or sets the pixel height of the element.</summary>
-        public double Height { get { return height; } }
+        public int Height { get { return height; } }
         #endregion
     }
 }
